moved frontend api call to helpers/methods
